chore(web-components): simplify compile script, tweak api-extractor config to stop spamming CI with warnings
